[UKET-46] 코틀린 프로젝트 모듈 합치기 (#1)
* chore : build.gradle 통합 및 settings.gradle에서 subproject 삭제

* file : 기존 모듈에 해당하는 코드들을 패키지로 이동

* chore : 누락된 의존성 추가

* style : lint 오류 해결
