+++ build.gradle.kts
@@ -8,19 +8,48 @@ plugins {
     id("org.jlleitschuh.gradle.ktlint") version "12.1.0"
 }
 
-allprojects {
-    group = "uket"
-    version = "0.0.1-SNAPSHOT"
+group = "uket"
+version = "0.0.1-SNAPSHOT"
 
-    repositories {
-        mavenCentral()
-    }
+repositories {
+    mavenCentral()
+}
 
-    tasks {
-        withType<Test> {
-            useJUnitPlatform()
-        }
-    }
+dependencies {
+    implementation("org.jetbrains.kotlin:kotlin-reflect")
+    implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+    implementation("org.springframework.boot:spring-boot-starter-web")
+
+    // test
+    testImplementation("org.springframework.boot:spring-boot-starter-test")
+    testImplementation("org.junit.jupiter:junit-jupiter:5.10.2")
+    testImplementation("org.assertj:assertj-core:3.25.3")
+    testImplementation("io.kotest:kotest-runner-junit5:5.8.0")
+    testImplementation("org.jeasy:easy-random-core:5.0.0")
+
+    // jwt
+    implementation("io.jsonwebtoken:jjwt-api:0.12.3")
+    implementation("io.jsonwebtoken:jjwt-impl:0.12.3")
+    implementation("io.jsonwebtoken:jjwt-jackson:0.12.3")
+
+    // swagger
+    implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0")
+
+    // security
+    implementation("org.springframework.boot:spring-boot-starter-security")
+    testImplementation("org.springframework.security:spring-security-test")
+
+    // DB
+    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+    runtimeOnly("com.mysql:mysql-connector-j")
+}
+
+tasks.test {
+    useJUnitPlatform()
+}
+
+kotlin {
+    jvmToolchain(21)
 }
 
 allOpen {
@@ -29,27 +58,6 @@ allOpen {
     annotation("javax.persistence.MappedSuperclass")
 }
 
-subprojects {
-    apply(plugin = "kotlin")
-    apply(plugin = "org.springframework.boot")
-    apply(plugin = "org.jetbrains.kotlin.plugin.spring")
-    apply(plugin = "io.spring.dependency-management")
-    apply(plugin = "org.jlleitschuh.gradle.ktlint")
-
-    kotlin {
-        jvmToolchain(21)
-    }
-
-    dependencies {
-        implementation("org.jetbrains.kotlin:kotlin-reflect")
-        implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-        testImplementation("org.junit.jupiter:junit-jupiter:5.10.2")
-        testImplementation("org.assertj:assertj-core:3.25.3")
-        testImplementation("io.kotest:kotest-runner-junit5:5.8.0")
-        testImplementation("org.jeasy:easy-random-core:5.0.0")
-    }
-
-    configure<KtlintExtension> {
-        verbose.set(true)
-    }
+configure<KtlintExtension> {
+    verbose.set(true)
 }
