+++ settings.gradle.kts
@@ -1,7 +1,1 @@
 rootProject.name = "uket"
-
-include("uket-api")
-include("uket-api:gateway")
-include("uket-api:infra")
-include("uket-api:application")
-include("uket-api:domain")
