+++ src/main/kotlin/application/port/TestPort.kt
@@ -1,0 +1,3 @@
+package uket.application.port
+
+interface TestPort
