+++ src/main/kotlin/application/usecase/TestUseCase.kt
@@ -1,0 +1,3 @@
+package uket.application.usecase
+
+interface TestUseCase
