+++ src/main/kotlin/domain/Test.kt
@@ -1,0 +1,3 @@
+package uket.domain
+
+class Test
