+++ src/main/kotlin/gateway/exception/BaseException.kt
