--- uket-api/application/src/main/kotlin/uket/application/port/TestPort.kt
@@ -1,3 +1,0 @@
-package uket.application.port
-
-interface TestPort {}
