--- uket-api/build.gradle.kts
@@ -1,6 +1,0 @@
-allprojects {
-    dependencies {
-        implementation("org.springframework.boot:spring-boot-starter-web")
-        testImplementation("org.springframework.boot:spring-boot-starter-test")
-    }
-}
