--- uket-api/domain/build.gradle.kts
@@ -1,1 +1,0 @@
-dependencies {}
