--- uket-api/gateway/build.gradle.kts
@@ -1,15 +1,0 @@
-dependencies {
-    implementation(project(":uket-api:infra"))
-
-    implementation("org.springframework.boot:spring-boot-starter-security")
-
-    // jwt
-    implementation("io.jsonwebtoken:jjwt-api:0.12.3")
-    implementation("io.jsonwebtoken:jjwt-impl:0.12.3")
-    implementation("io.jsonwebtoken:jjwt-jackson:0.12.3")
-
-    // swagger
-    implementation("org.springdoc:springdoc-openapi-starter-webmvc-ui:2.3.0")
-
-    testImplementation("org.springframework.security:spring-security-test")
-}
