--- uket-api/infra/build.gradle.kts
@@ -1,7 +1,0 @@
-dependencies {
-    implementation(project(":uket-api:domain"))
-    implementation(project(":uket-api:application"))
-
-    implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-    runtimeOnly("com.mysql:mysql-connector-j")
-}
